Fail early on any folder using a complex coder
By complex, any coder that takes and/or produces more than one stream of
bytes. BCJ2 is the only culprit to my knowledge.

I'm not convinced I'm navigating the bind pairs correctly at the moment
so the code panics with an out of bounds array access rather than error
that BCJ2 is unimplemented.